+++ scripts/actors/pc.js
@@ -86,6 +86,10 @@ export function o13pcActorMixin(base) {
 				}
 			}
 			
+			options.o13 = options.o13 ?? {};
+			
+			options.o13.shouldSyncImage = this.shouldSynchArchetypePortrait;
+			
 			await super._preUpdate(changed, options, user);
 		}
 		
@@ -97,6 +101,10 @@ export function o13pcActorMixin(base) {
 					if (changed.system.hasOwnProperty("archetype")) {
 						await this.updateArchetypeItems();
 						await this.synctoArchetypeBackground();
+						
+						if (options.o13?.shouldSyncImage) {
+							await this.synchArchetypePortrait();
+						}
 					}
 				}
 			}
@@ -112,7 +120,7 @@ export function o13pcActorMixin(base) {
 			}
 		}
 		
-		_onAROverrideChange(adddiff, remdiff) {
+		_onAEOverrideChange(adddiff, remdiff) {
 			if (adddiff.system?.hasOwnProperty("maxwounds") || remdiff.system?.hasOwnProperty("maxwounds")) {
 				this.updateMaxWounds();
 			}
@@ -164,6 +172,10 @@ export function o13pcActorMixin(base) {
 		//Story
 		get storyActor() {
 			return [...game.actors].find(actor => actor.isStory && actor.hasPC(this))
+		}
+		
+		get siblingCharacters() {
+			return this.storyActor?.pcActors?.filter(actor => actor != this) ?? [];
 		}
 		
 		//Acts
@@ -234,6 +246,8 @@ export function o13pcActorMixin(base) {
 		
 		async setOwnArchetype(archetype) {
 			if (archetype.isArchetype) {
+				const shouldSyncImage = this.shouldSynchArchetypePortrait;
+				
 				await this.removeOwnArchetype();
 				
 				await this.createEmbeddedDocuments("Item", [archetype.toObject()]);
@@ -252,6 +266,7 @@ export function o13pcActorMixin(base) {
 				else {
 					await this.updateArchetypeItems();
 					await this.synctoArchetypeBackground();
+					if (shouldSyncImage) await this.synchArchetypePortrait();
 				}
 			}
 		}
@@ -259,9 +274,12 @@ export function o13pcActorMixin(base) {
 		async removeOwnArchetype() {
 			const oldArchetypes = [...this.items].filter(item => item.isArchetype);
 				
+			const shouldSyncImage = this.shouldSynchArchetypePortrait;
+				
 			await this.deleteEmbeddedDocuments("Item", oldArchetypes.map(archetype => archetype.id));
 			
 			await this.updateArchetypeItems();
+			if (shouldSyncImage) await this.synchArchetypePortrait();
 		}
 		
 		get ownArchetype() {
@@ -284,6 +302,27 @@ export function o13pcActorMixin(base) {
 			if (archetype) {
 				return this.createEmbeddedDocuments("Item", [...Object.values(archetype.perksData), ...Object.values(archetype.guaranteedGear)]);
 			}
+		}
+		
+		async synchArchetypePortrait() {
+			const archetype = this.archetype;
+			
+			if (archetype?.img) {
+				if (this.img != archetype.img) {
+					return this.update({img : archetype.img})
+				}
+			}
+			else {
+				return this.update({img : this.defaultPortrait})
+			}
+		}
+		
+		get hasArchetypePortrait() {
+			return this.img == this.archetype?.img;
+		}
+		
+		get shouldSynchArchetypePortrait() {
+			return this.hasArchetypePortrait || this.hasDefaultPortrait;
 		}
 		
 		get archetypePrepState() {
@@ -330,6 +369,12 @@ export function o13pcActorMixin(base) {
 		usePerk(id) {
 			if (this.hasPickedPerk(id)) {
 				this.pickedPerks[id]?.use();
+			}
+		}
+		
+		restorePerkUse(id) {
+			if (this.hasPickedPerk(id)) {
+				this.pickedPerks[id]?.restoreUse();
 			}
 		}